separate mcp logics, handle mcp servers in agent
